audio_processing/requirements.txt
pydub
openai-whisper
git+https://github.com/pyannote/pyannote-audio.git
torch
torchvision
torchaudio
